requirements.txt
matplotlib==3.7.3
numpy==1.22.4
pandas==1.3.5
Pillow==6.2.1
Pillow==10.2.0
scikit_learn==1.2.0
scipy==1.12.0
seaborn==0.13.2
timm==0.6.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.64.1
